requirements.txt
# MCP Server dependencies
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# Vector embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Graph analysis
networkx>=3.0

# Authentication
bcrypt>=4.0.0
pyjwt>=2.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.4
httpx>=0.24.0